feat: 마이페이지 리뷰 api 코드 분리, /:id 경로 추가, 타입명 수정 및 user파일로 이동
