+++ frontend/src/components/mypage/PointAndReview.tsx
@@ -1,9 +1,9 @@
-import axios from "axios";
 import { useEffect, useState } from "react";
+import { useParams } from "react-router-dom";
 
 import { cn } from "@/lib/utils";
-import { Product } from "@/types/products";
-import { ReviewsData, ReviewUserData } from "@/types/reviews";
+import { getReview } from "@/services/user";
+import { MyReview } from "@/types/users";
 
 import {
   Tabs,
@@ -17,45 +17,18 @@ import ReviewList from "./ReviewList";
 
 export default function PointAndReviewTab() {
   const [tab, setTab] = useState(true);
-  const [reviews, setReviews] = useState<ReviewUserData[]>([]);
-
-  async function getReview() {
-    try {
-      const response = await axios.get<ReviewsData[]>(
-        "http://localhost:3000/reviews/user/1",
-      );
-      const myReviews: ReviewUserData[] = await Promise.all(
-        response.data.map(async (review) => {
-          return {
-            content: review.content,
-            //FIXME: model에서의 필드명: product_id, types에서의 속성명: productId 인데, types의 타입으로 db 반환 타입 지정하니 productId로 입력시 오류 -> types의 속성명 product_id로 수정해둠
-            productName: await getProductName(review.product_id),
-            date: review.createdAt.slice(0, 10),
-          };
-        }),
-      );
-      setReviews(myReviews);
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  async function getProductName(
-    productId: ReviewsData["product_id"],
-  ): Promise<string> {
-    try {
-      const response = await axios.get<Product>(
-        `http://localhost:3000/products/${productId}`,
-      );
-      return response.data.name;
-    } catch (error) {
-      console.log(error);
-      return "";
-    }
-  }
+  const [reviews, setReviews] = useState<MyReview[]>([]);
+  const {id} = useParams();
 
   useEffect(() => {
-    getReview();
+    async function fetchReviews() {
+      const myReview = await getReview(`${id}`);
+      if (myReview) {
+        setReviews(myReview);
+      }
+    }
+
+    fetchReviews();
   }, []);
 
   return (
