+++ frontend/src/types/users.ts
@@ -1,3 +1,5 @@
+import { ReviewsData } from "./reviews";
+
 export interface User {
   id?: number;
   name: string;
@@ -5,4 +7,11 @@ export interface User {
   phoneNumber: string; //
   password: string;
   adress: string;
+}
+
+export interface MyReview {
+  // id: Reviews['id'];
+  content: ReviewsData["content"];
+  productName: string;
+  date: string;
 }
